binance-telegram-bot/utils/market_analysis.py
import os
from binance.client import Client
from utils.indicators import *
import pandas as pd
from datetime import datetime
import random

binance_api_key = os.getenv("BINANCE_API_KEY")
binance_api_secret = os.getenv("BINANCE_API_SECRET")
client = Client(binance_api_key, binance_api_secret)

def analyze_market():
    symbols = ['BTCUSDT', 'ETHUSDT', 'SOLUSDT', 'BNBUSDT', 'AVAXUSDT']
    timeframe = '1h'
    limit = 100

    for symbol in symbols:
        klines = client.get_klines(symbol=symbol, interval=timeframe, limit=limit)
        df = pd.DataFrame(klines, columns=[
            'time','open','high','low','close','volume','close_time',
            'quote_asset_volume','number_of_trades','taker_buy_base_volume',
            'taker_buy_quote_volume','ignore'
        ])
        df['close'] = pd.to_numeric(df['close'])
        df = calculate_indicators(df)

        if is_rsi_oversold(df) and is_ema_crossover(df) and is_macd_bullish(df):
            return {
                'symbol': symbol.replace('USDT', '/USDT'),
                'side': 'BUY',
                'entry': float(df['close'].iloc[-1]),
                'tp1': round(float(df['close'].iloc[-1]) * 1.03, 2),
                'tp2': round(float(df['close'].iloc[-1]) * 1.05, 2),
                'sl': round(float(df['close'].iloc[-1]) * 0.97, 2),
                'volume_spike': round(random.uniform(2.5, 6.5), 1),
                'whale_activity': random.choice([True, False]),
                'news_impact': random.choice(['🟢 Positive', '🔴 Negative', '🟡 Neutral']),
                'sentiment': random.choice(['Bullish', 'Bearish', 'Neutral']),
                'trend_strength': random.choice(['Strong', 'Medium', 'Weak']),
                'timeframe': timeframe,
                'recommendation': 'STRONG BUY',
                'trade_type': random.choice(['Scalping', 'Normal', 'Spot']),
                'leverage': random.choice([20, 30, 50]),
                'signal_tag': f"Signal_{random.randint(1000,9999)}",
                'confidence': f"{round(random.uniform(95, 99.5), 2)}%"
            }
    return None
