binance-telegram-bot/requirements.txt
python-binance
telepot
requests
pandas
numpy
scikit-learn
textblob
schedule
